Create separate file for Book and Label's methods
